Update ClipboardModule.java (#251)

+++ android/src/main/java/com/reactnativecommunity/clipboard/ClipboardModule.java
@@ -238,7 +238,7 @@ public class ClipboardModule extends NativeClipboardModuleSpec {
   }
 
   @Override
-  public void removeListeners(int count) {
+  public void removeListeners(double count) {
 
   }
